Added Colored Bullets and Pistol
Changed bullet and pistol sprite to white

Bullet changes color by changing the color of the sprite renderer which adds a color mask over the object, changing it's color to the desired color.

Created a list of colors in the PlayerController script. Right now the list contains the default colors built into C# Unity but we can make our own colors by using  Color newColor = new Color(0.3f, 0.4f, 0.6f, 0.3f);
//uses rgba values for color

we need to discuss the art, more specifically the size of all of our sprites as that will determine certain settings for the camera, as well as the sizes of colliders and transforms. Ideally the transform sizes should all be 1,1,1 so the pixel art should be designed keeping that in mind.

the most important thing is that all of the sprites be the same resolution so that they can all be imported with the same settings.