Added ChangeColor script
Added the ChangeColor script. When attached to an object, if the object collides with a bullet, it's spriterenderer will change into the color of the bullet, or if the object is the same color as the bullet, it will print a statement indicating so. For the actual game this statement would be something that either lowers health or kills the player.