Jump Buffer working and added mirror
- Jump Buffer is working! Some of the code might be obsolete, should be tested by having debug statements for each if statement in the FixedUpdate method and seeing which if statements are actually being triggered and remove the ones that aren't.

- Added a Mirror Prefab with the "Mirror" Tag and "Ground" Layer. By putting the mirror on the ground layer, you are able to jump on mirrors. We should discuss whether or not we want the player to be able to jump on mirrors or not. This is a simple change as all it requires is changing the layer of the mirror prefab in the inspector from "Mirror" to "Untagged".

When bullets hit a mirror, they will bounce off instead of being destroyed. This is done using a Physics Material. The Physics Material is called "Bouncy Physics Material" and is located in the Physics Materials Folder. It has a friction of 0 and a bounciness value of 1 unlike the "Slippery Physics Material" that is attached to the player which has a friction value of 0  and a bounciness value of 0. In order to make the bullets bounce off of mirror, I had to turn off the IsTrigger value of the circle collider of the bullet prefab (this is done in the inspector by unchecking the box). Keep in mind that when writing code for detecting collisions with bullet you should use OnCollision instead of OnTrigger. Because the bullet colliders are no longer triggers, they now cause the player to move slightly off the ground when on the ground and shooting downwards. This should probably be fixed because it looks somewhat sloppy, however I think that trying to fix it would be more work than it would be worth.

- Working on getting controller input working. Single input buttons and left stick are working but right stick does not work because the code uses the mouses position on screen to aim rather than pointing in the vector2 direction that the stick is aiming at. Not sure about how I should approach this, should there be a different script for controller input? Or would it be better to use an if statement or something so that the way the direction of the gun is calculated changes based on whether you're using a mouse or controller. The most practical solution is to change it so that it works regardless of input device (controller or mouse) but I'm not sure how we would implement that because the right stick on a controller doesn't have a cursor that appears on screen so the ScreenToWorldPoint would not work.

- Testing the game in the game window is somewhat slow, so I recommend building and running the game if you want to test the game without making any changes while testing it.

- Currently the player can infinitely shoot bullets by clicking the left mouse button. We should begin implementing different weapons now as the basic player movement mechanics are mostly done at this point.