Made player sprite grey and some other notes (see description)
Made the player sprite grey by making the original sprite greyscale. This should be done by the artist with a more stylized look but I wanted to get something white/gray just for reference. The sprite needs to be white/grey so that when we change the color of the sprite renderer, the color's don't overlap. Also forgot to mention in the last push that I added a HideCursor script which hides the cursor on game start. If you click away from the game then go back in the mouse will be visible because it only hides the mouse on start instead of on update. We could choose to either keep the mouse visible or hide it, depends on group consensus. Currently working on getting right stick controller input working as the controller handles input differently from the mouse. They both pass Vector2's but the mouse passes it's position on screen while the controller passes the position of the stick on the controller (-1, -1) to (1, 1). Might have to make a new Action in the PlayerInputActions for this, not entirely sure, still figuring it out. Also I recommend checking out the Unity Beginner Playlist as I've added some more videos that are helpful and I will continue to add videos that I think are helpful/important to know. Link here https://youtube.com/playlist?list=PL9Vc8VwmDp3QRDCUCFiLCqN3Fs9Htaq65

Also working on adding particle effects so that when the bullets hit something they have a breaking animation or visual effect instead of just disappearing