Added Seamless Controller and Mouse support
You can now seamlessly transition between keyboard & mouse and controller. Added the Pseudo Mouse game object, which has a crosshair sprite that shows the position of the mouse and is the position that the player will aim at with a controller. The pseudo mouse can be moved using the right stick of a controller.

- cursor is now hidden using the HideCursor script because the pseudomouse will follow the mouse position if the boolean usingMouse = true;