Added Fast Falling
If the player holds downwards (S on keyboard, down on left stick of controller) they will fall faster to the ground. (Very similar to fast falling in Super Smash Bros. Ultimate). We should discuss whether or not we want this to be included. If not, it can be easily removed as it only requires a single if statement in the FixedUpdate method of the PlayerController.